http_push.c:  explicit cast computation of potential 32bit long to size_t
On Windows, computation may be performed as 32bit long rather than size_t.
Ensure all arguments are size_t to avoid implemenation dependent implicit
 conversion.

+++ http-push.c
@@ -370,7 +370,7 @@ static void start_put(struct transfer_request *request)
 
 	/* Set it up */
 	git_deflate_init(&stream, zlib_compression_level);
-	size = git_deflate_bound(&stream, len + hdrlen);
+	size = git_deflate_bound(&stream, len + (size_t) hdrlen);
 	strbuf_init(&request->buffer.buf, size);
 	request->buffer.posn = 0;
